requirements.txt
blinker>=1.8.2,<1.9.0
click==8.1.8
Flask==2.2.5
Flask-WTF==1.1.1
email-validator==1.3.1
flask-cors==5.0.1
Flask-Mail==0.9.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
Werkzeug==3.1.3
WTForms==3.2.1
python-dotenv==1.0.1